requirements.txt
tqdm==4.45.0
pillow==7.1.1
autopep8==1.5.1
numpy==1.18.2
fonttools==4.7.0
tensorflow==2.1.0
